pyproject.toml
[project]
name = "pdf-summarizer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit",
    "pypdf",
    "openai-agents>=0.6.1",
    "python-dotenv>=1.2.1",
    "tiktoken",
]
